requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
colorama==0.4.4
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djongo==1.3.6
dnspython==2.2.1
gunicorn==20.1.0
Pillow==9.0.1
PyJWT==2.3.0
pymongo==3.12.3
pytz==2021.3
qrcode==7.3.1
sqlparse==0.2.4
tzdata==2021.5
whitenoise==6.1.0
